changed to use DiscreteDynamicsWorld addRigidBody with collision masks when addRigidBody message instead a custom overlap filter callback using entity properties (the callback could be used for other purposes like dynamic masks or other properties, to check collisions).
git-svn-id: https://svn.gemserk.com/componentsengine/trunk/componentsengine@236 69e4d4fc-dda9-4998-a343-e80dfd0be058

+++ jbulletmodule/src/main/java/com/gemserk/componentsengine/jbullet/ComponentsEngineOverlapFilterCallback.java
@@ -34,14 +34,14 @@ public class ComponentsEngineOverlapFilterCallback extends OverlapFilterCallback
 		if (entity1 == null)
 			return collides;
 
-		Integer filterMask0 = Properties.getValue(entity0, "collisionFilterMask");
-		Integer filterGroup0 = Properties.getValue(entity0, "collisionFilterGroup");
+		Short filterMask0 = Properties.getValue(entity0, "collisionFilterMask");
+		Short filterGroup0 = Properties.getValue(entity0, "collisionFilterGroup");
 
 		if (filterMask0 == null || filterGroup0 == null)
 			return collides;
 
-		Integer filterMask1 = Properties.getValue(entity1, "collisionFilterMask");
-		Integer filterGroup1 = Properties.getValue(entity1, "collisionFilterGroup");
+		Short filterMask1 = Properties.getValue(entity1, "collisionFilterMask");
+		Short filterGroup1 = Properties.getValue(entity1, "collisionFilterGroup");
 
 		if (filterMask1 == null || filterGroup1 == null)
 			return collides;
